Retrigger when acquisition is complete.
In addition to retriggering, I also reduced the scope of lock(m_mutex)
to just the acquisition phase, just as it is done for the LeCroy scope.

This increases the responsiveness of the STOP button such that it reacts
immediately to the event. Without it, there are 4 or more AcquireData
calls before the STOP button event kicks in. During that time, the
LeCroyOscilloscope::Stop() is stalled at the lock(m_mutex) phase.