scopehal: Improve resource search on windows
Improve windows (mingw64) builds to look for resources such as
shaders, icons, etc. in the share directory adjacent to the bin
directory where execuatable lives. This matches the pattern for a
mingw64 package install, and is the same as on Linux.

Without this, a freshly installed mingw package will fail to start
the glscopeclient GUI due to inability to find resources.