USB2PacketDecoder: fix DecodeSetup behaviour when the ACK/NAK packet is missing
Previously, when looking for a response to a setup packet, DecodeSetup would
consume the subsequent packet whether it was an ACK/NACK or not.

In cases where the device did not reply, the decoder would consume a packet
from the next transaction and break the decoding.

This commit makes sure that DecodeSetup only consumes the subsequent packet
iff it is an ACK or NAK.

fixes #723