RigolOscilloscope: fix DS1000Z max download chunk size (maxpoints)
Initial tests with 1 channel and 1 MB chunks worked fine.
Later tests with 2 channels failed, but 500 kB chunk size worked fine,
but failed for 3 or 4 channels. It looks like the chunk size (maxpoints),
also depends on the amount of enabled channels.
There may be still some margin, but it's not worth chaseing it.