RigolOscilloscope: add experimental support for MSO5xx2 4CH indirect option detection
MSO5xx2 scopes are scopes with only 2 channels and other 2 channels are SW option.
Usually `:SYST:OPT:STAT? 4CH` query should be enough, but there are modifier FWs
which report `0` (not present), but the option is active.
This is and attempt to detect it using `:CHAN3:VERN ON`,
which is expected to trigger error if option is missing.
We intentionally do not use any CHAN3 query, as those time out.